separate protocol-name-list
