+++ catalystwan/models/policy/__init__.py
@@ -25,6 +25,7 @@ from catalystwan.models.policy.list.policer import PolicerList
 from catalystwan.models.policy.list.port import PortList
 from catalystwan.models.policy.list.preferred_color_group import PreferredColorGroupList
 from catalystwan.models.policy.list.prefix import PrefixList
+from catalystwan.models.policy.list.protocol_name import ProtocolNameList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -41,17 +42,7 @@ from .definition.security_group import SecurityGroup
 from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
-from .lists import (
-    ProtocolNameList,
-    RegionList,
-    SiteList,
-    SLAClassList,
-    TLOCList,
-    URLAllowList,
-    URLBlockList,
-    VPNList,
-    ZoneList,
-)
+from .lists import RegionList, SiteList, SLAClassList, TLOCList, URLAllowList, URLBlockList, VPNList, ZoneList
 from .lists_entries import PathPreference, PolicerExceedAction
 from .localized import LocalizedPolicy
 from .policy_definition import (
